removed unecessary crypto and cli_utils functions and moved to wallet_utils
